Some file renames and moving class definitions out of single files.
