+++ NIHR.ProfileManagement/NIHR.ProfileManagement.Api/Models/Dto/CreateUserRequestDto.cs
@@ -8,8 +8,4 @@
         public string lastname { get; set; }
 
     }
-
-    public class CreateUserResponseDto
-    {
-    }
 }
