+++ NIHR.ProfileManagement/NIHR.ProfileManagement.Domain/Models/CreateProfileResponse.cs
@@ -1,0 +1,8 @@
+﻿
+namespace NIHR.ProfileManagement.Domain.Models
+{
+    public class CreateProfileResponse
+    {
+        public ProfileInfo? Profile { get; set; }
+    }
+}
